+++ module/item/item-attack.js
@@ -66,7 +66,7 @@ export async function AttackOptions(item) {
         const combatants = game?.combat?.combatants;
         if (combatants && typeof dragRuler != 'undefined') {
 
-            if (tokens.length === 1) {
+            if (token) {
 
                 let distance = dragRuler.getMovedDistanceFromToken(token);
                 let speed = dragRuler.getRangesFromSpeedProvider(token)[1].range;
@@ -80,7 +80,7 @@ export async function AttackOptions(item) {
         }
 
         // Simplistic velocity calc using dragRuler
-        if (data.velocity === 0) {
+        if (data.velocity === 0 && token) {
             if (typeof dragRuler != 'undefined') {
                 if (dragRuler.getRangesFromSpeedProvider(token).length > 1) {
                     data.velocity = parseInt(dragRuler.getRangesFromSpeedProvider(token)[1].range || 0);